+++ frontend/components/EmptyCart/index.jsx
@@ -1,21 +1,38 @@
 import React from 'react';
 import I18n from '@shopgate/pwa-common/components/I18n';
+import { themeConfig } from '@shopgate/engage';
+import PropTypes from 'prop-types';
 import Icon from './components/Icon';
 import styles from './style';
 
+const { svgImages = {} } = themeConfig || {};
+const { emptyCart = '' } = svgImages || {};
+
 /**
  * The Cart Empty component.
+ * If configured, theme config will override the icon here
+ * @param {React.node} children children to display if image is configured in theme config
  * @return {JSX.Element}
  */
-const Empty = () => (
+const Empty = ({ children }) => (
   <div className={styles.container}>
-    <div className={styles.icon}>
-      <Icon />
+    <div className={emptyCart ? null : styles.icon} aria-hidden>
+      {emptyCart ? children : <Icon />}
     </div>
-    <div className={styles.title}>
-      <I18n.Text string="cart.empty" />
-    </div>
+    {!emptyCart && (
+      <div className={styles.title}>
+        <I18n.Text string="cart.empty" />
+      </div>
+    )}
   </div>
 );
+
+Empty.propTypes = {
+  children: PropTypes.node,
+};
+
+Empty.defaultProps = {
+  children: null,
+};
 
 export default Empty;