+++ frontend/components/ProductSlider/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { useTheme } from '@shopgate/engage/core';
 import { I18n, ButtonLink } from '@shopgate/engage/components';
+import { ProductSlider as BaseProductSlider } from '@shopgate/engage/product/components';
 import getConfig from '../../helpers/getConfig';
 import connect from './connector';
 import styles from './style';
@@ -18,7 +18,7 @@ const { showOnPdpPage, showOnEmptyCartPage } = getConfig();
  * @param {string} showMoreUrl Path to show more page
  * @param {string} headline Headline for Product slider
  * @param {string} className className
- * @return {JSX}
+ * @return {JSX.Element}
  */
 const ProductSlider = ({
   isCartPage,
@@ -41,8 +41,6 @@ const ProductSlider = ({
     return null;
   }
 
-  const { ProductSlider: BaseProductSlider } = useTheme();
-
   let defaultHeadline = 'recently_viewed_products.headline';
 
   if (isCartPage) {
@@ -54,7 +52,7 @@ const ProductSlider = ({
   return (
     <div className={`${styles.slider} ${className}`}>
       <div className={styles.headlineContainer}>
-        <h3 className={styles.headline(hasShowMore)}>
+        <h3 className={`${styles.headline(hasShowMore)} recently-viewed-products__product-slider__headline`}>
           <I18n.Text string={headline || defaultHeadline} />
         </h3>
         {hasShowMore && (